% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae}
\alias{ae}
\title{Adverse event data}
\format{
\subsection{\code{ae}}{

A data frame with 46 rows and 25 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\usage{
ae
}
\description{
Adverse event data
}
\keyword{datasets}
