% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalstat-package.R
\docType{package}
\name{clinicalstat-package}
\alias{clinicalstat}
\alias{clinicalstat-package}
\title{clinicalstat: Clinical Statistics Analysis}
\description{
Clinical Statistics Analysis.
}
\author{
\strong{Maintainer}: Xinlong Yang \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
