% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_empty2.R
\name{zap_empty2}
\alias{zap_empty2}
\title{replace empty strings with NA}
\usage{
zap_empty2(.data, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{specific columns names of \code{.data},If ignored, select all columns.}
}
\value{
A data frame
}
\description{
\code{zap_empty2()} allows you to replace empty strings in all columns or specific columns
of a data frame with NA.
}
\examples{

t <- zap_empty2(ae,AETERM)

t <- zap_empty2(ae,AETERM,AEREL)

t <- zap_empty2(ae)
}
